AutoScroll RichTextBox only if viewing its end
Previously, when AutoScroll was activated, the RichTextBox was scrolled
to its end every time a log message was added. This made it very difficult
to browse previous log messages as new ones were being appended.

Added a check to scroll to the end only if the RichTextBox was already
scrolled to the end.